Module cansat_stm32f4::app

The RTIC application module

Re-exports

pub use rtic::Monotonic as _;

Modules

blink

Software task

gps_irq

Hardware task

idle

Idle loop

init

Initialization function

monotonics

Holds static methods for each monotonic.

send_meas

Software task

shared_resources 🔒

Structs

Local 🔒

Shared 🔒

__rtic_internal_Monotonics

Monotonics used by the system

__rtic_internal_blinkLocalResources

Local resources blink has access to

__rtic_internal_blink_Context

Execution context

__rtic_internal_blink_MicrosecMono_SpawnHandle

__rtic_internal_gps_irqSharedResources

Shared resources gps_irq has access to

__rtic_internal_gps_irq_Context

Execution context

__rtic_internal_idleLocalResources

Local resources idle has access to

__rtic_internal_idleSharedResources

Shared resources idle has access to

__rtic_internal_idle_Context

Execution context

__rtic_internal_initLocalResources

Local resources init has access to

__rtic_internal_init_Context

Execution context

__rtic_internal_send_measLocalResources

Local resources send_meas has access to

__rtic_internal_send_measSharedResources

Shared resources send_meas has access to

__rtic_internal_send_meas_Context

Execution context

__rtic_internal_send_meas_MicrosecMono_SpawnHandle

Constants

CEILING 🔒

Priority ceiling

CEILING 🔒

Priority ceiling

CHECK 🔒

PRIORITY 🔒

PRIORITY 🔒

The priority of this interrupt handler

Statics

DEFMT_LOG_STATEMENT 🔒

Functions

EXTI0 🔒 ⚠

Interrupt handler to dispatch tasks at priority 1

TIM2 🔒 ⚠

USART1 🔒 ⚠

__rtic_internal_blink_MicrosecMono_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_blink_MicrosecMono_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_blink_spawn

Spawns the task directly

__rtic_internal_send_meas_MicrosecMono_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_send_meas_MicrosecMono_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_send_meas_spawn

Spawns the task directly

check 🔒

idle 🔒

init 🔒

User code end

Type Definitions

MicrosecMono 🔒

User code from within the module